Fixed some functionality not working under Windows XP
The stat() function is broken under VS2015 when using the v140_xp
platform toolset. This caused .raw files to be taken as HDD images,
with confusing consequences. Thanks to Gorski for reporting this.

Added FileSize() helper and Win32-specific implementations of
other functions that use stat. Also removed some conditional code
that called the appropriate version of stat, which was messy.
When MS finally fix the bug the new helpers can be simplified.

+++ include/SAMdisk.h
@@ -198,7 +198,7 @@ typedef void * HANDLE;
 #include "Util.h"
 #include "SAMCoupe.h"
 
-static const size_t MAX_IMAGE_SIZE = 64 * 1024 * 1024;	// 64MiB
+static const int MAX_IMAGE_SIZE = 64 * 1024 * 1024;	// 64MiB
 
 enum { GAPS_AUTO = -1, GAPS_NONE, GAPS_CLEAN, GAPS_TRACK, GAPS_ALL };
